Logic for determining which holidays occur on each day
This provides a mechanism to determine which holidays occur on a
specific day.  A standard set of US centric holidays have been included
along with some more excentric days of celebration and commemoration.

Two mechanisms are used to determine the dates.  First a set of fixed
date holidays is included.  Additionally, a set of variable holidays
that fall on different days each year, typically determined by the day
of the week is also included.